Create Rogue - Trapsmith [M.T. Black's work, transcribed by Darren].js
My first attempts at scripts for MPMB's character sheet.
The description for the Booby Traps ability is *very* long and at high level the abilities will need to be moved to the notes section. I don't know if it is possible to have the script send some of the description to the notes section on it's own, but that would be ideal.